requirements.txt
# DeployMind - Multi-Agent Autonomous Deployment System
# All packages are free/open-source

# AI Framework
crewai>=0.28.0
anthropic>=0.18.0

# Web Framework (for future API)
fastapi>=0.109.0
uvicorn>=0.27.0

# AWS
boto3>=1.34.0

# GitHub
PyGithub>=2.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Cache
redis>=5.0.0

# Docker
docker>=7.0.0

# CLI
click>=8.1.0
rich>=13.7.0

# Security Scanning
# Trivy is installed separately (system binary)

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Logging
python-json-logger>=2.0.0

# Utilities
pydantic>=2.5.0
httpx>=0.26.0
